Refactor Args out.
