+++ tetwild.h
@@ -13,8 +13,11 @@
 
 #include <array>
 #include <vector>
+#include "Args.h"
 
 namespace tetwild {
+
+    extern Args parameters;
 
     void tetrahedralization(const std::vector<std::array<double, 3>>& V_in,
                             const std::vector<std::array<int, 3>>& F_in,
